moved model to external file
