+++ model/Thought.js
@@ -1,0 +1,19 @@
+import mongoose from 'mongoose'
+
+const Thought = mongoose.model('Thought', {
+  text: {
+    type: String,
+    required: true,
+    minlength: 5,
+  },
+  like: {
+    type: Boolean,
+    default: false,
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
+})
+
+export default Thought
